// src/client/scheduling/shared/p-scheduling-calendar/p-shift-item-module/shift-item/shift-item-styles.ts
export enum ShiftItemViewStyles {
	SMALL = 'small',
	DETAILED = 'detailed',
	BUTTON = 'button',
	MEDIUM = 'medium',
	MEDIUM_MULTI_SELECT = 'mediumMultiSelect',
	MULTI_SELECT = 'multiSelect',
}

export enum PCalendarShiftStyle {
	OVERVIEW = 'overview',
	SHIFT_PICKER = 'shiftPicker',
	FULL = 'full',
}